requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
flask_login==0.6.3
flask_wtf==1.2.1
flask_migrate==4.0.7
email_validator==2.2.0
psycopg2-binary==2.9.9
gunicorn==22.0.0

